+++ Scripts/Items/Addons/LoomEastAddon.cs
@@ -43,9 +43,9 @@ namespace Server.Items
         {
             base.Serialize(writer);
 
-            writer.Write((int)1); // version
+            writer.Write(1); // version
 
-            writer.Write((int)m_Phase);
+            writer.Write(m_Phase);
         }
 
         public override void Deserialize(GenericReader reader)
@@ -95,7 +95,7 @@ namespace Server.Items
         {
             base.Serialize(writer);
 
-            writer.Write((int)0); // version
+            writer.Write(0); // version
         }
 
         public override void Deserialize(GenericReader reader)
