+++ Scripts/Items/Artifacts/Equipment/Armor/DefenderOfTheMagus.cs
@@ -118,7 +118,7 @@ namespace Server.Items
         {
             base.Serialize(writer);
 
-            writer.Write((int)0);//version
+            writer.Write(0);//version
         }
     }
 }
