+++ Scripts/Items/Artifacts/Equipment/Armor/HumanSignOfOrder.cs
@@ -79,7 +79,7 @@
         public override void Serialize(GenericWriter writer)
         {
             base.Serialize(writer);
-            writer.Write((int)0); //version
+            writer.Write(0); //version
         }
     }
 }
