+++ Scripts/Items/Artifacts/Equipment/Clothing/RunedSashOfWarding.cs
@@ -122,7 +122,7 @@ namespace Server.Items
         public override void Serialize(GenericWriter writer)
         {
             base.Serialize(writer);
-            writer.Write((int)0);
+            writer.Write(0);
             writer.Write(m_Charges);
             writer.Write((int)m_Ward);
         }
@@ -156,7 +156,7 @@ namespace Server.Items
         {
             base.Serialize(writer);
 
-            writer.Write((int)0); // version
+            writer.Write(0); // version
         }
 
         public override void Deserialize(GenericReader reader)
